Use cls parameter in NGram class methods
Use the cls class parameter passed to the NGram class methods. This
parameter is the only reason to mark something a class method, and it
should be used if so marked.

This change also provides a small but consistent speed-up due to using a
local rather than a global variable, with the normalize function on a
critical path.