[Fix multithreading problem]
Profiles are now loaded as soon as the object is created.

Before, this was done in two steps, creating a race condition on the
global variable _factory: as soon as the line '_factory =
DetectorFactory()' was executed, the other threads jumped the condition
'if _factory is None' going through 'create' and '_create_detector',
until they met the line 'if not self.langlist'.

Since the thread that created the global DetectorFactory didn't
necessarily have the time to populate the language profiles, the other
threads start raising LangDetectException until the langlist is no
longer empty (which is not good either: it can have just one language...).

Since the global variable is used to avoid reloading the profiles again
and again (which can take some time), I'm leaving it global, but the
profile loading process must therefore be done as soon as the object is
created, so when the other threads detect that '_factory' is loaded,
it's really going to be loaded (not partially loaded). :)